backend/requirements.txt
flask 
flask-restful 
flask-cors 
requests 
python-dotenv